pyproject.toml
[project]
name = "pipedream"
version = "0.1.0"
description = "Real-time generative visualization for interactive fiction."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.47.0",
    "google-generativeai>=0.8.6",
    "litellm>=1.80.16",
    "opencv-python>=4.12.0.88",
    "pexpect>=4.9.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipedream"]
